requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
black==22.12.0
click==8.1.3
Django==4.1.5
djangorestframework==3.14.0
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.11.0
platformdirs==2.6.2
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
